# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAPhoto API
  description: API for WASAPhoto, a webapp for sharing photos.
  version: 3.1.0
servers:
  - url: http://localhost:3000
    description: For development purposes
security:
  - UsernameAuth: [] 
paths:
  /session:
    post:
      tags: ["login"]
      summary: Logs in the user
      description: |-
        If the user does not exist, it will be created,
        and an identifier is returned.
        If the user exists, the user identifier is returned.
      security: []
      operationId: doLogin
      requestBody:
        description: User details
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  $ref: "#/components/schemas/Username"
        required: true
      
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  identifier:
                    $ref: "#/components/schemas/UniqueID"
        
        '200':
          description: User already exists
          
        '400': { $ref: "#/components/responses/BadRequest" }
        '500': { $ref: "#/components/responses/InternalServerError" }
...
  /users/{userId}/photo:
    parameters:
      - $ref: "#/components/parameters/UserID"
      
    get:
      parameters:
        - $ref: "#/components/parameters/Page"
      tags: ["photos"]
      summary: Get a list of user photos
      description: Retrieve photos in reverse chronological order
      operationId: getUserPhotos
      
      responses:
        '200':
          description: A list of photos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Photos"
        
        '400': { $ref: "#/components/responses/BadRequest" }
        '401': { $ref: "#/components/responses/Unauthorized" }
        '403': { $ref: "#/components/responses/Forbidden" }
        '500': { $ref: "#/components/responses/InternalServerError" }
      
      
    post:
      tags: ["my photos"]
      summary: Upload a photo
      description: | 
        Upload a photo for the authenticated user. Once submitted the server uploads the photo
        to a dedicated server/website and attaches a new id, author's username, date, time, 
        and initiates likesAmount and commentsAmount to 0.
        
      operationId: uploadPhoto
      requestBody:
        content:
          image/*:
            schema:
              $ref: "#/components/schemas/RawPhoto"
        required: true

      responses:
        '201':
          description: Photo uploaded successfully
        
        '400': { $ref: "#/components/responses/BadRequest" }
        '401': { $ref: "#/components/responses/Unauthorized" }
        '403': { $ref: "#/components/responses/Forbidden" }
        '500': { $ref: "#/components/responses/InternalServerError" }
  /users/{userId}/photo/{photoId}:
    parameters:
      - $ref: "#/components/parameters/UserID"
      - $ref: "#/components/parameters/PhotoID"
      
    delete:
      tags: ["my photos"]
      summary: Remove a photo
      description: Remove a specific photo for the authenticated user.
      operationId: deletePhoto
      
      responses:
        '200':
          description: Photo removed successfully
        
        '204':
          description: Photo not found
        
        '400': { $ref: "#/components/responses/BadRequest" }
        '401': { $ref: "#/components/responses/Unauthorized" }
        '403': { $ref: "#/components/responses/Forbidden" }
        '500': { $ref: "#/components/responses/InternalServerError" }
  /users/{userId}/followers:
    parameters:
      - $ref: "#/components/parameters/UserID"
      
    get:
      parameters:
        - $ref: "#/components/parameters/Page"
      tags: ["followers"]
      summary: Get followings
      description: |
        Get a list of followers(userIds) of a user in order to see any
        particular user's details use dedicted path /users/{userId}
      operationId: getFollowers
      
      responses:
        '200':
          description: List of user profiles 
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Users"
          
        '400': { $ref: "#/components/responses/BadRequest" }
        '401': { $ref: "#/components/responses/Unauthorized" }
        '500': { $ref: "#/components/responses/InternalServerError" }
  /users/{userId}/followings:
    parameters:
      - $ref: "#/components/parameters/UserID"
      
    get:
      parameters:
        - $ref: "#/components/parameters/Page"
      tags: ["followings"]
      summary: Get followings
      description: |
        Get a list of followings(userIds) of a user in order to see any
        particular user's details use dedicted path /users/{userId}
      operationId: getFollowings
      
      responses:
        '200':
          description: List of user profiles 
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Users"
          
        '400': { $ref: "#/components/responses/BadRequest" }
        '401': { $ref: "#/components/responses/Unauthorized" }
        '500': { $ref: "#/components/responses/InternalServerError" }
  /users/{userId}/followings/{otherUserId}:
    parameters:
      - $ref: "#/components/parameters/UserID"
      - $ref: "#/components/parameters/OtherUserID"
      
    put:
      tags: ["followings"]
      summary: Follow a user
      description: Follow a specific user 
      operationId: followUser
  
      responses:
        '201':
          description: User followed successfully
        
        '200':
          description: User already followed
        
        '400': { $ref: "#/components/responses/BadRequest" }
        '401': { $ref: "#/components/responses/Unauthorized" }
        '403': { $ref: "#/components/responses/Forbidden" }
        '500': { $ref: "#/components/responses/InternalServerError" }
      
    delete:
      tags: ["followings"]
      summary: Unfollow a user
      description: Unfollow a specific user by his/her username
      operationId: unfollowUser
      
      responses:
        '200':
          description: User unfollowed successfully
        
        '400': { $ref: "#/components/responses/BadRequest" }
        '401': { $ref: "#/components/responses/Unauthorized" }
        '403': { $ref: "#/components/responses/Forbidden" }
        '500': { $ref: "#/components/responses/InternalServerError" }
  /users/{userId}/banned:
    parameters:
      - $ref: "#/components/parameters/UserID"
    
    get:
      parameters:
        - $ref: "#/components/parameters/Page"
      tags: ["ban"]
      summary: Get banned users
      description: |
        Get a list of banned users(userIds) of a user in order to see any
        particular user's details use dedicted path /users/{userId}
      operationId: getBannedUsers
      
      responses:
        '200':
          description: List of user profiles 
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Users"
          
        '400': { $ref: "#/components/responses/BadRequest" }
        '401': { $ref: "#/components/responses/Unauthorized" }
        '500': { $ref: "#/components/responses/InternalServerError" }
  /users/{userId}/banned/{otherUserId}:
    parameters:
      - $ref: "#/components/parameters/UserID"
      - $ref: "#/components/parameters/OtherUserID"
      
    put:
      tags: ["ban"]
      summary: Ban a user
      description: Ban a specific user 
      operationId: banUser
      
      responses:
        '201':
          description: User banned successfully
        
        '200':
          description: User already banned
          
        '400': { $ref: "#/components/responses/BadRequest" }
        '401': { $ref: "#/components/responses/Unauthorized" }
        '403': { $ref: "#/components/responses/Forbidden" }
        '500': { $ref: "#/components/responses/InternalServerError" }
      
    delete:
      tags: ["ban"]
      summary: Unban a user
      description: Unban a specific user by his/her username
      operationId: unbanUser
      
      responses:
        '200':
          description: User unbanned successfully
          
        '400': { $ref: "#/components/responses/BadRequest" }
        '401': { $ref: "#/components/responses/Unauthorized" }
        '403': { $ref: "#/components/responses/Forbidden" }
        '500': { $ref: "#/components/responses/InternalServerError" }
...
  /photos/{photoId}/comment:
    parameters:
      - $ref: "#/components/parameters/PhotoID"

    get:
      tags: ["comments"]
      summary: Get comments
      description: Get a list of all comments under the photo
      operationId: getComments
      
      responses:
        '200': 
          description: A list of comments
          content:
            application/json:
              schema: 
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
                  
        '400': { $ref: "#/components/responses/BadRequest" }
        '401': { $ref: "#/components/responses/Unauthorized" }
        '500': { $ref: "#/components/responses/InternalServerError" }
  
    post:
      tags: ["comments"]
      summary: Add a comment to a photo
      description: Add a comment to a specific photo.
      operationId: commentPhoto
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Comment"
              
      responses:
        '201':
          description: Comment added successfully
          
        '400': { $ref: "#/components/responses/BadRequest" }
        '401': { $ref: "#/components/responses/Unauthorized" }
        '500': { $ref: "#/components/responses/InternalServerError" }
...
components:
  securitySchemes:
    UsernameAuth:
      type: http
      scheme: bearer
  parameters:
    PhotoID:
      schema:
        $ref: "#/components/schemas/UniqueID"
      name: photoId
      in: path
      required: true
      description: Photo ID
...
    UserID:
      schema:
        $ref: "#/components/schemas/UniqueID"
      name: userId
      in: path
      required: true
      description: User ID
    OtherUserID:
      schema:
        $ref: "#/components/schemas/UniqueID"
      name: otherUserId
      in: path
      required: true
      description: User ID
...
    Page:
      name: page
      in: query
      required: false
      description: |
        The page number for paginating through the data. Each page contains a set number of items of interest.
      schema:
        type: integer
        minimum: 1
        default: 1
        example: 1
  responses:
    BadRequest:
      description: The request was not compliant with the documentation (eg. missing fields, etc)
    Unauthorized:
      description: The access token is missing or invalid
    Forbidden:
      description: You do not have access rights to the operation
    InternalServerError:
      description: The server encountered an internal error. Further info in server logs
  schemas:
    UniqueID:
      description: UUID for any object needed
      type: string
      pattern: '^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$'
      minLength: 36
      maxLength: 36
      readOnly: true
      example: "550e8400-e29b-41d4-a716-446655440000"
    Username:
      type: string
      example: Maria
      pattern: 'ˆ.*?$'
      minLength: 3
      maxLength: 32 
    RawPhoto:
      description: binary of an image
      type: object
      properties:
        photoBinary:
          type: string
          minLength: 0
          format: binary
          example: "010010101"
    Timestamp:
      description: Date and time representation
      type: string
      format: date-time
      example: "2023-01-01T12:00:00Z"
      minLength: 20
      maxLength: 20
      readOnly: true
...
    Users:
      description: Paginated list of users
      type: array
      minItems: 1
      maxItems: 25
      items:
        type: object
        properties:
          userId:
            $ref: "#/components/schemas/User/properties/userID"
          username:
            $ref: "#/components/schemas/User/properties/username"
...
    Photos:
      description: |
        paginated list of photo ids from which you can retrieve 
        a particular photo using /photos/{photoID} endpoint
      type: array
      minItems: 1
      maxItems: 10
      items:
        $ref: "#/components/schemas/Photo/properties/photoID"
    Comment:
      description: Representation of a single comment
      type: object
      properties:
        commentID:
          $ref: "#/components/schemas/UniqueID"
        author: 
          $ref: "#/components/schemas/UniqueID"
        timestamp:
          $ref: "#/components/schemas/Timestamp"
        commentText:
          type: string
          example: "You look good, my man :)"
      required:
        - commentText  
...